svn:3786 Nubar retrieval changed. The nubar retrieval routine (in utils) should be called after empire is executed to retrieve the evaluated nubar data from /data/nubar.endf file if the file NUBAR.DAT is not available in the local directory.
Default OMP deformation changed back to Moller-Nix resulting in a more realistic assignment of deformed or spherical nuclei. Note that Nobre parameterization gives the dynamical deformation, which is very good for vibrational calculations, but not to define whether the nucleus is deformed. E.g. Fe-56 becomes deformed with Nobre param., while it is very soft, but spherical in its ground state with small dynamical deformation as predited by Moller-Nix tables.