svn:4045 Modify parsing of input file to allow for free-formatting of the value and the 4 optional integer qualifiers. The parameter name is still restricted to maximum 6 characters, and the value is expected to be entered, but defaults to 0.0.
The integer qualifiers are not required and default to 0. This should
be backwardly compatible with the old "G10.5,4I5" format except for
comments at the end of the line. The old fixed-format allowed for
comments in columns beyond the last integer. With free-format, there
needed some way to signify comments, so I chose the "!" character to
indicate that from that character to the end of the line as comments
and not to be parsed for parameter values or qualifiers. So, if your
input file does not start comments with the "!" character, you'll need
to add this to your file before using this free format version. The
values themselves are delimited by the space " " character.

Files added: parse_line.f90
Files modified: input.f, main.f, and the Makefile

secis06.f was modifed here only by the cosmetic removal of trailing
ctrl_M characters by running the file through dos2unix.