svn:4387 Redefinition of the meaning of the ENDF with respect to setting exclusive spectra.
ENDF = 0 no advanced formatting and no exclusive spectra (no change)

ENDF = 1 advanced ENDF formatting with exclusive spectra for first 4 emitted neutrons and first proton. This is the RECOMMENDED OPTION for creating ENDF file.

ENDF > 1 declares as exclusive a rectangle of nuclei of ENDF protons per ENDF neutrons with compound nucleus in the upper right corner. EMPEND will try to use this info in formatting but, in general, not all results can be disantangled and formatted resulting in the inconsistent ENDF file. NOT to be used for ENDF file production!

Logic of the exclusive flag assignment greatly simplified and win2unix run.