svn:4542 HMS updated to keep good balance and fix the fluctuations in DDXS. Legendre polynomial fit for integration was abandoned, a simple sum is used for the integration. The calculated DDXS are fixed to avoid drops in the DDXS distribution by extrapolating forward and backward from calculated "big" values.
In the test calculation the balance is perfect and the DDXS look reasonable. I expect EMPEND to succeed in the Legendre fit of those data.

The new empend is crashing :-), more work is needed.

The gamma spectrum for (z,d)=(z,np+pn+d) has been split into (z,d)
and (z,np+pn) with different Q-values. That should improve the energy balance of MT=28 and MT=105. However, empend processing is still outstanding.