svn:5118 WFC() allocated inside the loop after all elastic and outgoing channels collected only for Moldauer.
Fission commented in EW as it does not work properly. In fact the treatment of uncoupled channels is still shaky. But the effect for U-238 is big at 0.26 MeV. At that energy fission and capture remains more or less the same but elastic decreases by 50 mb  if EW is active !