svn:5151 According to the last version of ENDF-6 manual (Chapter 40, 40.4 item 3.b) when MF1=MF, the XMF1 field shall be entered as blank or zero. Therefore, within  a given subsection, the sub-subsections for MF1=MF will always appear before those for other MF1 values.
The same happens for the error related to LB=8. In the ENDF-6 manual (Chapter 40, items 40.2 & 40.3) it is written that the formats of the sub-sub-subsection for File 40 are exactly the same as the formats for the sub-subsections for File 33 and the procedures for File 40 are the same as for File 33 except that File 40 has one more level of indexing corresponding to the LFS and LFS1 flags.

Therefore, Daniel Lopez patched checkr.f (see !DLA comments).