svn:5322 New version of OPTMAN sent by Roberto Capote to Matteo Vorabbi on an email sent on January 22, 2022. Roberto's message is below:
" Dave and Matteo,
Dmitry Martyanov fixed and tested a new version. Hopefully all issues that you have identified are fixed. Please check before I commit the changes,
We still have a problem with parallel execution but this is only needed for fitting.
In EMPIRE we use OPTMAN only for one energy at a time.
Best regards
Roberto"

From the files sent, I had to preserve the previous Makefile and also make a change to OPTMAND.f which was to include the following in the first line:

#define EMPMODE