svn:5416 (i) Change of selection exclusive residues (again!), (ii) forgoten changes related to reading common part of input.
(i) ENDF=1 selects pure neutron and pure proton residues as exclusive; ENDF=2 selects 2x2 rectangle (as before); ENDF=n selects nxn rectangle (not recommended). In edition there is new input keyword "EXCLUS" which allows to eliminate exclusive flag from any nucleus in the rectangle or add the flag to any nucleus outside the rectangle. In the latter case it is up to a user to ensure that all pathes to the nucleus are exclusive. This change allows to eliminate nuclei which are populated by different reactions and maintain only those which have a clear single path. Small contamination might be tollerated since EMPEND will eliminate spectra that do not belong to the reaction (e.g., n, p and d from (n,a)). Unfortunately, this can be done only manually by checking exclusive spectra in .lst for the highest incident energy.